Rename k8stools module to k8s
