+++ strimziregistryoperator/k8s.py
